+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerDirector.java
@@ -125,7 +125,7 @@ public class TigerDirector {
                         .build());
                     try {
                         await().pollInterval(1, TimeUnit.SECONDS)
-                            .atMost(5, TimeUnit.HOURS)
+                            .atMost(getLibConfig().getPauseExecutionTimeoutSeconds(), TimeUnit.SECONDS)
                             .until(() -> tigerTestEnvMgr.isUserAcknowledgedShutdown());
                     } finally {
                         tigerTestEnvMgr.shutDown();
@@ -341,7 +341,7 @@ public class TigerDirector {
                 .bannerIsHtml(isHtml)
                 .build());
             await().pollInterval(1, TimeUnit.SECONDS)
-                .atMost(5, TimeUnit.HOURS)
+                .atMost(getLibConfig().getPauseExecutionTimeoutSeconds(), TimeUnit.SECONDS)
                 .until(() -> tigerTestEnvMgr.isUserAcknowledgedContinueTestRun());
             tigerTestEnvMgr.resetUserInput();
         } else {
@@ -363,7 +363,7 @@ public class TigerDirector {
                 .bannerType(BannerType.FAIL_PASS)
                 .build());
             await().pollInterval(1, TimeUnit.SECONDS)
-                .atMost(5, TimeUnit.HOURS)
+                .atMost(getLibConfig().getPauseExecutionTimeoutSeconds(), TimeUnit.SECONDS)
                 .until(() -> tigerTestEnvMgr.isUserAcknowledgedContinueTestRun()
                     || tigerTestEnvMgr.isUserAcknowledgedFailingTestRun());
             if (tigerTestEnvMgr.isUserAcknowledgedFailingTestRun()) {