+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerLibConfig.java
@@ -27,4 +27,6 @@ public class TigerLibConfig {
     public boolean activateWorkflowUi = false;
     @Builder.Default
     public boolean createRbelHtmlReports = true;
+    @Builder.Default
+    public long pauseExecutionTimeoutSeconds = 18000L;
 }